bri.f90: Prevent underflow of UTEN and USTAR by setting to zero when small. bri.f90: Moved declaration of T1 to globals.f90 to preserve contents. vel.f90: Moved declaration of T1 to globals.f90 to preserve contents. vegrad.f90: Moved declaration of T1 to globals.f90 to preserve contents. globals.f90: Added variables T1, CR, OK, RC as global declarations. mom.f90: Removed "save" declaration requirement by moving CR, OK and RC to globals.f90,    and by initializing some variables upon re-entry each time. config_gen.f90: Updated default values of cloud_flag, cld_frac and index_soils. tests.f90, test_bri.f90, test_netrad.f90, test_load_data.f90: Updated to operate with above changes.

+++ src/config_mod.f90
@@ -23,7 +23,7 @@ module config_mod
 
   type t_met
     real(kind=real64) :: omega, zo, obst_hgt
-    integer :: cld_fract
+    real(kind=real64) :: cld_fract
     logical :: cloud_flag
   end type t_met
